size_t for memzero and nonzero

+++ common/memzero.c
@@ -1,6 +1,6 @@
 #include "../common.h"
 
-void memzero(void* a, unsigned long n)
+void memzero(void* a, size_t n)
 {
 	char* p = (char*) a;
 	char* e = p + n;
